fix(events): refactor Pagure events
