+++ tests/unit/test_checkers.py
@@ -52,6 +52,7 @@ from packit_service.worker.checker.vm_image import (
 from packit_service.worker.events import (
     AbstractCoprBuildEvent,
     PullRequestGithubEvent,
+    PushPagureEvent,
 )
 from packit_service.worker.events.event import EventData
 from packit_service.worker.events.github.pr import (
@@ -59,7 +60,6 @@ from packit_service.worker.events.github.pr import (
 )
 from packit_service.worker.events.github.push import Push as PushGitHubEvent
 from packit_service.worker.events.gitlab import MergeRequestGitlabEvent, PushGitlabEvent
-from packit_service.worker.events.pagure import PushPagureEvent
 from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
 from packit_service.worker.mixin import ConfigFromEventMixin
